backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
Flask-JWT-Extended>=4.5
Flask-Cors>=4.0
requests>=2.28
APScheduler>=3.10
pytz
openai
groq
requests
supabase
playwright>=1.40.0
PyPDF2>=3.0.0
langchain
langchain-groq